v1.2 - Removed deprecated UI placement calls
Due to wxPython updates:
Specifiying horitzonal placements in horizontal boxes (e.g. right\left) throw an error instead of a warning.
This also applies to vertical boxes with top/bottom flags.
I removed the flags. The UI looks normal from tests so far and the software was able to run.